MCPClient: use both clamscan and clamdscan
Added new `CLAMAV_CLIENT_THRESHOLD` setting to MCPClient in order to let the
user define a threshold in MB so when the size of a file being scanned for
viruses is too big the scanner can fall back to clamscan because clamd can't
handle big files not even when the file is passed by reference.

This closes #753.